+++ price-adapter-raw/src/lib.rs
@@ -66,8 +66,10 @@ mod binance_websocket;
 mod coingecko;
 mod coinmarketcap;
 pub mod error;
+mod huobipro;
 pub mod types;
 
 pub use binance_websocket::{BinanceWebsocket, BinanceWebsocketService};
 pub use coingecko::CoinGecko;
 pub use coinmarketcap::CoinMarketCap;
+pub use huobipro::HuobiPro;