+++ price-adapter-raw/CHANGELOG.md
@@ -4,4 +4,5 @@
 
 - Support CoinGecko sources
 - Support Binance Websocket source
+- Support Huobipro sources
 - Support CoinMarketCap source