+++ price-adapter/CHANGELOG.md
@@ -5,4 +5,5 @@
 - Support CoinGecko sources
 - Support CoinMarketCap source
 - Support Binance Websocket source
+- Support HuobiPro sources
 - Create Interval and Websocket service to manage sources